adjust aprsidaily

+++ sh/aprsidaily.sh
@@ -13,9 +13,10 @@ fi
 DBuser=$(echo    `grep '^DBuser '   $CONFIGDIR/APRSconfig.ini` | sed 's/=//g' | sed 's/^DBuser //g')
 DBpasswd=$(echo  `grep '^DBpasswd ' $CONFIGDIR/APRSconfig.ini` | sed 's/=//g' | sed 's/^DBpasswd //g' | sed 's/ //g' )
 DBpath=$(echo    `grep '^DBpath '   $CONFIGDIR/APRSconfig.ini` | sed 's/=//g' | sed 's/^DBpath //g' | sed 's/ //g' )
+DBname=$(echo    `grep '^DBname '   $CONFIGDIR/APRSconfig.ini` | sed 's/=//g' | sed 's/^DBname //g' | sed 's/ //g' )
 SCRIPT=$(readlink -f $0)
 SCRIPTPATH=`dirname $SCRIPT`
-SARpath=$(echo    `grep '^DBpath '   $CONFIGDIR/SARconfig.ini` | sed 's/=//g' | sed 's/^DBpath //g' | sed 's/ //g' )
+SARpath=$(echo    `grep '^DBpath '  $CONFIGDIR/SARconfig.ini` | sed 's/=//g'  | sed 's/^DBpath //g' | sed 's/ //g' )
 #SARpath='/nfs/OGN/DIRdata/'
 
 
@@ -72,21 +73,23 @@ elif
    sudo rm /tmp/tempfile* heat*    										     >/dev/null     2>/dev/null
 fi
 date														     >>APRSproc.log 2>/dev/null
-echo "clean OGNDATA in APRSLOG"							                                     >>APRSproc.log 2>/dev/null
-echo "DELETE FROM RECEIVERS WHERE otime < date('"$(date +%Y-%m-%d)"')-3;" | mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG >>APRSproc.log 2>/dev/null
-echo "INSERT INTO OGNDATAARCHIVE SELECT * FROM OGNDATA;      " | mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG     >>APRSproc.log 2>/dev/null
-echo "DELETE FROM OGNDATAARCHIVE where date < '"$date"'; "     | mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG     >>APRSproc.log 2>/dev/null
-echo "DELETE FROM OGNTRKSTATUS WHERE otime < date('"$(date +%Y-%m-%d)"'); " | mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG >>APRSproc.log 2>/dev/null
-echo "SELECT COUNT(*) from OGNTRKSTATUS  ; "                   | mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG     >>APRSproc.log 2>/dev/null
-echo "DELETE FROM OGNDATA;"                                    | mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG     >>APRSproc.log 2>/dev/null
-echo "DELETE FROM GLIDERS ; "                                  | mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG     >>APRSproc.log 2>/dev/null
-echo "INSERT INTO GLIDERS  SELECT * FROM OGNDB.GLIDERS;      " | mysql -u $DBuser -p$DBpasswd    -h $server APRSLOG     >>APRSproc.log 2>/dev/null
-echo "SELECT COUNT(*) from GLIDERS  ; "                        | mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG     >>APRSproc.log 2>/dev/null
-echo "Report number of GLIDERS with the LAST FIX POSITION"			                                     >>APRSproc.log 2>/dev/null
-echo "SELECT COUNT(*) from GLIDERS_POSITIONS   ; "             | mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG     >>APRSproc.log 2>/dev/null
-echo "DELETE from GLIDERS_POSITIONS WHERE length(flarmId) < 6;"| mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG     >>APRSproc.log 2>/dev/null
-echo "DELETE from GLIDERS_POSITIONS WHERE flarmId like '%RND%' ;"| mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG   >>APRSproc.log 2>/dev/null
-echo "SELECT COUNT(*) from GLIDERS_POSITIONS   ; "             | mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG     >>APRSproc.log 2>/dev/null
+if [[ $DBname == 'APRSLOG' ]]; then
+	echo "clean OGNDATA in APRSLOG"							                                     >>APRSproc.log 2>/dev/null
+	echo "DELETE FROM RECEIVERS WHERE otime < date('"$(date +%Y-%m-%d)"')-3;" | mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG >>APRSproc.log 2>/dev/null
+	echo "INSERT INTO OGNDATAARCHIVE SELECT * FROM OGNDATA;      " | mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG     >>APRSproc.log 2>/dev/null
+	echo "DELETE FROM OGNDATAARCHIVE where date < '"$date"'; "     | mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG     >>APRSproc.log 2>/dev/null
+	echo "DELETE FROM OGNTRKSTATUS WHERE otime < date('"$(date +%Y-%m-%d)"'); " | mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG >>APRSproc.log 2>/dev/null
+	echo "SELECT COUNT(*) from OGNTRKSTATUS  ; "                   | mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG     >>APRSproc.log 2>/dev/null
+	echo "DELETE FROM OGNDATA;"                                    | mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG     >>APRSproc.log 2>/dev/null
+	echo "DELETE FROM GLIDERS ; "                                  | mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG     >>APRSproc.log 2>/dev/null
+	echo "INSERT INTO GLIDERS  SELECT * FROM OGNDB.GLIDERS;      " | mysql -u $DBuser -p$DBpasswd    -h $server APRSLOG     >>APRSproc.log 2>/dev/null
+	echo "SELECT COUNT(*) from GLIDERS  ; "                        | mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG     >>APRSproc.log 2>/dev/null
+	echo "Report number of GLIDERS with the LAST FIX POSITION"			                                     >>APRSproc.log 2>/dev/null
+	echo "SELECT COUNT(*) from GLIDERS_POSITIONS   ; "             | mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG     >>APRSproc.log 2>/dev/null
+	echo "DELETE from GLIDERS_POSITIONS WHERE length(flarmId) < 6;"| mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG     >>APRSproc.log 2>/dev/null
+	echo "DELETE from GLIDERS_POSITIONS WHERE flarmId like '%RND%' ;"| mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG   >>APRSproc.log 2>/dev/null
+	echo "SELECT COUNT(*) from GLIDERS_POSITIONS   ; "             | mysql -u $DBuser -p$DBpasswd -v -h $server APRSLOG     >>APRSproc.log 2>/dev/null
+fi
 date														     >>APRSproc.log 2>/dev/null
 if [ -d /var/www/html/files ]
 then
@@ -100,13 +103,13 @@ then
 fi
 if [[ $(hostname) == 'CHILEOGN' ]]
 then
-   mysqldump -u $DBuser -p$DBpasswd -h $server --add-drop-table APRSLOG TRKDEVICES  >/var/www/html/files/TRKDEVICES.sql 2>/dev/null
+	mysqldump -u $DBuser -p$DBpasswd -h $server --add-drop-table APRSLOG TRKDEVICES  >/var/www/html/files/TRKDEVICES.sql 2>/dev/null
 else
-   if [[ -f TRKDEVICES.sql ]]
-   then
-      rm TRKDEVICES.sql												     >/dev/null 2>/dev/null
-   fi
-   wget chileogn.ddns.net/files/TRKDEVICES.sql -o /tmp/TRKDEVICES.sql						     >/dev/null 2>/dev/null
+   	if [[ -f TRKDEVICES.sql ]]
+   	then
+      		rm TRKDEVICES.sql												     >/dev/null 2>/dev/null
+   	fi
+   	wget chileogn.ddns.net/files/TRKDEVICES.sql -o /tmp/TRKDEVICES.sql						     >/dev/null 2>/dev/null
 fi
 if [[ -f TRKDEVICES.sql && $(hostname) != 'CHILEOGN' ]]
 then
@@ -119,7 +122,7 @@ else
         pt-table-sync  --execute --verbose h=chileogn.ddns.net,D=APRSLOG,t=TRKDEVICES h=$server --user=$DBuser --password=$DBpasswd >>APRSproc.log 2>/dev/null
 fi
 rm /tmp/TRKDEVICES.sql
-echo "Done."     		     						                                     >>APRSproc.log 2>/dev/null
+echo "Done."$(date)    		     						                                     >>APRSproc.log 2>/dev/null
 date														     >>APRSproc.log 2>/dev/null
 mutt -a APRSproc.log -s $hostname" APRSlog daily report ..." -- $(cat $SCRIPTPATH/mailnames.txt)		     2>/dev/null 
 mv APRSproc.log  archive/APRSPROC$(date +%y%m%d).log 								     2>/dev/null
